third_party/WebKit/LayoutTests/fast/css/invalidation/sub-selector-adjacent-cancellation-expected.txt

Negated selector should not clear adjacent combinator flag.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getComputedStyle(b).backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 2
PASS getComputedStyle(b).backgroundColor is green
PASS successfullyParsed is true

TEST COMPLETE
